Putting the core::execute implementation into a whole new file to enhance compile time
