+++ src/cpu/Rules.mk
@@ -6,7 +6,7 @@ src/cpu/$(OBJECT_DIR)/Core.o: src/cpu/Core.cpp src/cpu/Core.hpp \
   src/Gameboy.hpp src/AMemoryBankController.hpp src/Debugger.hpp \
   src/MemoryBus.hpp src/Cartridge.hpp src/MemoryBankController1.hpp \
   src/PPU.class.h src/ScreenOutput.class.h \
-  src/cpu/InterruptController.hpp src/cpu/instructions.inc
+  src/cpu/InterruptController.hpp
 src/cpu/Core.hpp:
 src/Fwd.hpp:
 src/IReadWrite.hpp:
@@ -20,6 +20,9 @@ src/MemoryBankController1.hpp:
 src/PPU.class.h:
 src/ScreenOutput.class.h:
 src/cpu/InterruptController.hpp:
+src/cpu/$(OBJECT_DIR)/CoreExecute.o: src/cpu/CoreExecute.cpp \
+  src/cpu/Core.hpp src/Fwd.hpp src/IReadWrite.hpp \
+  utils/Operations_utils.hpp src/cpu/instructions.inc
 src/cpu/instructions.inc:
 src/cpu/$(OBJECT_DIR)/InterruptController.o: \
   src/cpu/InterruptController.cpp src/cpu/InterruptController.hpp \
