+++ src/cpu/Sources.mk
@@ -1,6 +1,7 @@
 ## Auto generated make tool, don't edit manually.
 
 OBJ_FILES_src/cpu := Core.o \
+                     CoreExecute.o \
                      InterruptController.o \
                      Timer.o
